requirements.txt
gradio
pandas
mysql-connector-python
sqlparse
python-dotenv
groq
